@@ -60,68 +60,79 @@
     "uuid-encoder": "^1.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.6.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.7.4",
-    "@contentful/rich-text-react-renderer": "^13.2.0",
+    "@babel/core": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@babel/register": "^7.8.3",
     "@storybook/addon-a11y": "~5.2.8",
     "@storybook/addon-actions": "~5.2.8",
     "@storybook/addon-info": "~5.1.11",
     "@storybook/addon-knobs": "~5.2.8",
     "@storybook/react": "~5.2.8",
-    "@types/jest": "^24.0.22",
-    "@types/react": "^16.9.14",
-    "@types/react-dom": "^16.9.4",
-    "@types/styled-components": "^4.4.0",
+    "@testing-library/jest-dom": "^5.0.2",
+    "@testing-library/react": "^9.4.0",
+    "@testing-library/react-hooks": "^3.2.1",
+    "@types/jest": "^25.1.0",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
+    "@types/storybook__addon-info": "^5.2.1",
+    "@types/styled-components": "^4.4.2",
     "@types/styled-system": "^5.1.4",
-    "@typescript-eslint/eslint-plugin": "^2.8.0",
-    "@typescript-eslint/parser": "^2.8.0",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
-    "babel-plugin-remove-graphql-queries": "^2.7.10",
-    "babel-plugin-styled-components": "^1.10.6",
+    "babel-plugin-remove-graphql-queries": "^2.7.22",
+    "babel-plugin-styled-components": "^1.10.7",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-preset-react-app": "^9.1.0",
-    "codecov": "^3.2.0",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "enzyme-to-json": "^3.4.0",
-    "eslint": "^6.7.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-config-react-app": "^5.0.2",
-    "eslint-plugin-flowtype": "^4.3.0",
-    "eslint-plugin-import": "^2.18.2",
+    "codecov": "^3.6.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
+    "enzyme-to-json": "^3.4.3",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-config-react-app": "^5.1.0",
+    "eslint-plugin-flowtype": "^4.6.0",
+    "eslint-plugin-import": "^2.20.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.18.0",
     "eslint-plugin-react-hooks": "^2.3.0",
     "eslint-plugin-tsc": "^1.2.0",
-    "gatsby-plugin-typescript": "^2.1.15",
-    "husky": "^3.1.0",
-    "jest": "^24.9.0",
-    "jest-environment-jsdom": "^24.0.0",
-    "jest-styled-components": "^6.3.1",
-    "jest-transform-stub": "^2.0.0",
-    "lint-staged": "^9.4.3",
-    "prettier": "^1.18.2",
-    "prop-types": "^15.6.2",
-    "react": "^16.9.0",
+    "gatsby-plugin-typescript": "^2.1.26",
+    "husky": "^4.2.1",
+    "jest": "^25.1.0",
+    "jest-styled-components": "^6.3.4",
+    "lint-staged": "^10.0.3",
+    "prettier": "^1.19.1",
+    "prop-types": "^15.7.2",
     "react-docgen-typescript-loader": "^3.6.0",
-    "react-dom": "^16.12.0",
     "remark-cli": "^7.0.1",
     "remark-preset-lint-recommended": "^3.0.3",
-    "stylelint": "^12.0.0",
+    "stylelint": "^13.0.0",
     "stylelint-a11y": "^1.2.2",
-    "stylelint-config-prettier": "^6.0.0",
+    "stylelint-config-prettier": "^8.0.1",
     "stylelint-config-recommended": "^3.0.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-prettier": "^1.1.1",
+    "stylelint-prettier": "^1.1.2",
     "stylelint-processor-styled-components": "^1.9.0",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.2",
+    "typescript": "^3.7.5",
     "typescript-styled-plugin": "^0.15.0"
+  },
+  "resolutions": {
+    "@reach/router": "^1.1.1",
+    "eslint-plugin-flowtype": "^3.0.0",
+    "eslint-plugin-react-hooks": "^1.0.0",
+    "graphql": "^14.1.1",
+    "jquery": "^3.4.1",
+    "react-side-effect": "^2.1.0",
+    "react-test-renderer": "^16.9.0",
+    "react-with-direction": "^1.1.0",
+    "regenerator-runtime": "^0.13.3",
+    "stylelint": "^11.0.0",
+    "webpack": "^4.0.0"
   }
 }
