@@ -3,4 +3,7 @@ declare module '*.svg' {
   export default content
 }
 
-declare module '*.jpg'
+declare module '*.jpg' {
+  const content: any
+  export default content
+}
