@@ -6,9 +6,9 @@ module.exports = {
     '\\.(css|jpg|png|svg)$': '<rootDir>/empty-module.js',
   },
   setupFiles: ['<rootDir>/config/jest.setup.js'],
+  setupFilesAfterEnv: ['<rootDir>/config/jest.setup.ts'],
   snapshotSerializers: ['enzyme-to-json/serializer'],
   coverageDirectory: './coverage/',
-  collectCoverage: true,
   globals: {
     __PATH_PREFIX__: '',
   },
